% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_curve.R
\name{add_curve}
\alias{add_curve}
\title{Add a fitted linear model to a plot}
\usage{
add_curve(mod, col = "black", plot_base = trellis.last.object())
}
\arguments{
\item{mod}{Model. A linear model with a single response and a single explanatory
variable.}

\item{col}{Character. The name of a color to use when plotting the line/curve.}
}
\description{
Include a line or curve created from a linear model with a single response
and a single explanatory variable. The \code{mod} function should be created
by using the \code{lm} function. To plot a polynomial curve, use the
\code{poly} function with the explanatory variable. See the examples below
for additional usage details.
}
\examples{
# Create a linear model and use the model plot the line on a xyplot
m_line <- lm(weight ~ height, data = cdc)
xyplot(weight ~ height, data = cdc)
add_curve(m_line)

# Create a linear model with a quadratic polynomial and include the curve on
# the xyplot.
m_curve <- lm(weight ~ poly(height, 2), data = cdc)
xyplot(weight ~ height, data = cdc)
add_curve(m_curve)

}
\seealso{
\code{\link{add_line}}, \code{\link{lm}}, \code{\link{poly}}
}
