% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{arm_span}
\alias{arm_span}
\title{Arm Span and Height Data}
\format{
A data frame with 90 observations of 4 variables.
}
\usage{
data(arm_span)
}
\description{
A dataset containing the arm spans, heights and additional information from a
group of Los Angeles high school students.
}
\details{
\itemize{ \item age. The age of the student. \item gender_label. The gender
of the student. \item armspan. The self-reported length of the student's arm
span, in inches. \item height. The self-reported height of the student, in
inches. }
}
\keyword{dataset}
