% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{atu_clean}
\alias{atu_clean}
\title{American Time Use Survey Data Sample - Clean}
\format{
A data frame with 10,493 observations of 8 variables
}
\source{
\url{http://www.bls.gov/tus/}
}
\usage{
data(atu_clean)
}
\description{
A dataset containing a subset of variables from the American Time Use Survey.
This dataset is a cleaned version of \code{atu_dirty}.
}
\details{
\itemize{ \item caseid. unique identifier of individual survey participant
\item age. the age of the respondent \item gender. the gender of the
respondent  \item fulltime_emp. the employment status of the respondent \item
phys_challenge. does the respondent have a physical difficulty  \item sleep.
the length of time the person sleeps, in minutes \item homework. How long the
respondent spent on homework assignments, in minutes \item socializing. the
number of minutes the respondent spent socializing }
}
\seealso{
\link{atu_dirty}
}
\keyword{dataset}
