% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{cereal_test}
\alias{cereal_test}
\title{Cereal test data}
\format{
A data frame with 20 observations of 12 variables
}
\usage{
data(cereal_test)
}
\description{
A dataset about the nutritional value of different breakfast cereals used to test models created with the cereal data.
}
\details{
\itemize{
\item cereal. The name of the cereal.
\item mfr. The manufacturer of the cereal.
\item calories. The number of calories contained in a single serving.
\item protein. The amount of protein contained in a single serving.
\item fat. The amount of fat contained in a single serving.
\item sodium. The amount of sodium contained in a single serving.
\item fibre. The amount of fibre contained in a single serving.
\item carbo. The amount of carbohydrates contained in a single serving.
\item sugars. The amount of sugar contained in a single serving.
\item shelf. The number of shelfs above the floor the cereal should be displayed on.
\item potassium. The number of  contained in a single serving.
\item vitamins. The type of vitamins included in the cereal.
}
}
\seealso{
\link{cereal}
}
\keyword{dataset}
