% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{colorize}
\alias{colorize}
\title{Add colors by category}
\usage{
colorize(x)
}
\arguments{
\item{x}{Categorical. Used to base the coloring of the observations on. For
now must have 12 or fewer categories to work.}
}
\value{
When used with \code{mutate}, the function returns a variable where
  each observation is assigned a color based on the categories of a
  categorical variable.
}
\description{
A simple function to add a variable of colors to a data frame based on a
separate categorical variable. Designed to be used within the \code{dplyr}
\code{mutate} funciton.
}
\examples{
mt_colors <- mutate(mtcars, colors = colorize(cyl))
head(mt_colors$colors)
xyplot(hp ~ disp, data = mt_colors, groups = colors)

}
