% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{Load built-in datasets.}
\usage{
data(...)
}
\arguments{
\item{...}{The name of the dataset to load.}
}
\value{
Loads the dataset and opens it up to view.
}
\description{
Load built-in data sets. After loading the data, the data frame will popup in
an RStudio window to view.
}
\examples{
data(cdc)
}
