% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{dwp_teacher}
\alias{dwp_teacher}
\title{LA Department of Water and Power (2010-2011) Data - Teacher}
\format{
A data frame with 512 observations of 19 variables
}
\source{
\url{https://www.ladwp.com}
}
\usage{
data(dwp_teacher)
}
\description{
Water usage from the Los Angeles Department of Water and Power, by census
block, for July 2010 through June 2011. Some blocks are dropped due to identifiability
concerns. This data set is used for a student practicum and contains 512 observations can be used by teachers to validate the student models.
}
\details{
\itemize{ \item census. The census block. \item sector_type. Indicates the
type of customers using water. \item longitude. The longitudinal centroid of
the census block. \item latitude. The latitudinal centroid of the census
block. \item census_pop. The number of people residing within the census
block. \item total. The total amount of water used in the year in hundreds of
cubic feet for each census block. \item july 2010 - june 2011. The amount of water used
per month in hundreds of cubic feet for each census block. \item count. The
number of users measured per census block by sector type.}
}
\seealso{
\link{dwp_2010}, \link{dwp_student}
}
\keyword{dataset}
