% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{extra_data}
\alias{extra_data}
\title{Load external datasets.}
\usage{
extra_data(data)
}
\arguments{
\item{data}{The name of the dataset to load.}
}
\value{
Loads the dataset and opens it up to view.
}
\description{
Load external data sets. After loading the data, the data frame will popup in
an RStudio window to view. For more detail, check curriculum.idsucla.org/data
}
\examples{
extra_data('cdc')
}
