% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lab.R
\name{extra_labs}
\alias{extra_labs}
\alias{extra_lab}
\title{Load new extra lab slides into RStudio's Viewer Pane}
\usage{
extra_labs(lab)

extra_lab(lab)
}
\arguments{
\item{lab}{Integer/String (optional). Including an integer will load the desired extra lab
automatically. Leaving the argument blank will load a menu for users to
make a selection from.}
}
\description{
Load new extra lab slides into RStudio's Viewer Pane. The function can either be called
without an argument \code{extra_lab()}, in which case a menu will appear in
the console for the user to make their selection, or with a number indicating
the desired lab as an argument.
}
\examples{
extra_lab() # Loads a menu to choose from
extra_lab(999) # Automatically loads the 999th lab from the menu.
extra_lab('veryinterestinglab') # Automatically loads the veryinterestinglab from the menu.
extra_labs() # Similar to extra_lab()
extra_labs(999) # Similar to extra_lab(999)
extra_labs('veryinterestinglab') # Similar to extra_lab('veryinterestinglab')

}
