% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{fev}
\alias{fev}
\title{Lung capacity data}
\format{
A data frame with 654 observations of 4 variables
}
\source{
\url{http://ww2.amstat.org/publications/jse/v13n2/datasets.kahn.html}
}
\usage{
data(fev)
}
\description{
A dataset containing a subset of variables from the American Time Use Survey.
This dataset is "dirty", meaning it has elements which require formatting
before use.
}
\details{
\itemize{ \item age. Age in years
\item lung_cap. Maximum forced expiratory volume in one second \item height. Height in inchces \item gender. "Female" or "Male" \item smoker. Describes whether a person is a smoker ("Yes") or not ("No")}
}
\keyword{dataset}
