% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{flights}
\alias{flights}
\title{March 2020 Flight Data}
\format{
A data frame with 310 observations of 11 variables
}
\source{
\url{https://aspm.faa.gov}
}
\usage{
data(flights)
}
\description{
This dataset contains the number of flights and the number of passengers at
each of the top 10 busiest airports in the U.S. in March 2020.
}
\details{
\itemize{
\item date.
\item departures. Number of departure flights.
\item arrivals. Number of arrival flights.
\item total_operations. Number of departure and arrival flights.
\item departure_seats. Number of available seats on departure flights.
\item arrival seats. Number of available seats on arrival flights
\item average_departure_seats. Average number of seats on a departure flight.
\item average_arrival_seats. Average number of seats on a arrival flight.
\item airport. IATA Airport Code
\item day. Day of the week, starting with Sunday.
\item week. Week of the Month}
}
\keyword{dataset}
