% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line.R
\name{get_line}
\alias{get_line}
\title{Calculate Slope-Intercept Form of a Line}
\usage{
get_line(p1, p2)
}
\arguments{
\item{p1}{The coordinates of the first point, specified as a numeric vector of length 2.}

\item{p2}{The coordinates of the second point, specified as a numeric vector of length 2.}
}
\value{
A named numeric vector containing the intercept and slope of the line.
}
\description{
Given two points, calculates the slope-intercept form of a line.
}
\examples{
p1 <- c(0, 0)
p2 <- c(1, 1)
get_line(p1, p2)
# Output: intercept = 0, slope = 1
}
