% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{head}
\alias{head}
\title{Print first six observations.}
\usage{
head(x, ..., data = NULL, groups = NULL, na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{Formula or other object to print.}

\item{data}{Data frame. The data where the variables can be found.}
}
\description{
\code{head} will print the first six values of an object.
}
\examples{
head(~hp, data = mtcars)
tail(~hp, data = mtcars)

}
