% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kclusters.R
\name{kclusters_simple}
\alias{kclusters_simple}
\title{Simple k-Means for 2 variable clustering}
\usage{
kclusters_simple(x, y, k, ...)
}
\arguments{
\item{x}{Formula. An expression of the form \code{y ~ x} where \code{x} and
\code{y} are the variables to use for clustering.}

\item{k}{Integer. The number of clusters to group the data into.}
}
\description{
A utilization of the \code{kmeans} function to simplify k-means clustering of
2 variables.
}
\examples{
# Use k-means to cluster football/soccer players by height and weight.
data(futbol)
kclusters(wt_lbs ~ ht_inches, data = futbol, k = 2)

# Include this data in the original data and plot it
futbol <- mutate(futbol, cluster = kclusters(wt_lbs ~ ht_inches, data = futbol, k = 2))
xyplot(wt_lbs ~ ht_inches, data = futbol, groups = cluster)

# We can compare the clustering to the actual
xyplot(wt_lbs ~ ht_inches, data = futbol, groups = league)

}
\seealso{
\code{\link{stats::kmeans}}
}
