% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lab.R
\name{load_lab}
\alias{load_lab}
\alias{load_labs}
\title{Load lab slides into RStudio's Viewer Pane}
\usage{
load_lab(lab)

load_labs(lab)
}
\arguments{
\item{lab}{Integer (optional). Including an integer will load the desired lab
automatically. Leaving the argument blank will load a menu for users to
make a selection from.}
}
\description{
Load lab slides into RStudio's Viewer Pane. The function can either be called
without an argument \code{load_lab()}, in which case a menu will appear in
the console for the user to make their selection, or with a number indicating
the desired lab as an argument.
}
\examples{
load_lab() # Loads a menu to choose from
load_lab(1) # Automatically loads the 1st lab from the menu.
load_labs() # Similar to load_lab()
load_labs(1) # Similar to load_lab(1)

}
