% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lab.R
\name{load_pds}
\alias{load_pds}
\alias{load_pd}
\title{Load new pd lab slides into RStudio's Viewer Pane}
\usage{
load_pds(lab)

load_pd(lab)
}
\arguments{
\item{lab}{Integer (optional). Including an integer will load the desired pd lab
automatically. Leaving the argument blank will load a menu for users to
make a selection from.}
}
\description{
Load new pd lab slides into RStudio's Viewer Pane. The function can either be called
without an argument \code{load_pd()}, in which case a menu will appear in
the console for the user to make their selection, or with a number indicating
the desired lab as an argument.
}
\examples{
load_pd() # Loads a menu to choose from
load_pd(1) # Automatically loads the 1st lab from the menu.
load_pds() # Similar to load_pd()
load_pds(1) # Similar to load_pd(1)

}
