% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{data}
\name{log_config}
\alias{log_config}
\title{A set of logging functions that Steve made.
These are and aren't used by the user.}
\format{
An object of class \code{logger} of length 3.
}
\usage{
log_config
}
\description{
A set of logging functions that Steve made.
These are and aren't used by the user.
}
\keyword{datasets}
