% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Make a map of points}
\usage{
map(
  x,
  data = parent.frame(),
  e,
  scaleby,
  size = 1,
  col = "blue",
  zoom = 10,
  color,
  ...
)
}
\arguments{
\item{x}{Function of the form latitude~longitude}

\item{data}{Data frame containing latitude and longitude variables}

\item{scaleby}{A variable for bubbles to be scaled by in a bubble plot.}

\item{size}{Size of the points. Default is 1, larger numbers will make points larger.}

\item{col}{A specification for the plotting color. Defaults to "blue". You
can provide any color name here, like "red", "purple", etc., in quotes. Or,
you could use hex values like "FF00FF", again in quotes. See section 'Color
Specification' in the \code{\link{par}} documentation for more.}

\item{dots}{graphical parameters to be passed to \code{\link{points}}. For
example, you could change the size of the dots on the map by passing
\code{cex=2}.}
}
\value{
A map of points with a google map background
}
\description{
A function that creates a plot of points on a background map pulled from
Google Maps' API.
}
\details{
Creates a map using a Google Maps basemap, with points of data on top. Points
can be colored differently and sized according to another variable
}
\examples{
lat <- runif(10, 32, 42)
lon <- runif(10, -124, -114)
geodata <- data.frame(lat, lon)
scale <- sample(1:10, 10)
map(lat~lon, data=geodata)
map(lat~lon, data=geodata, scaleby = scale)
map(lat~lon, data=geodata, scaleby = scale, size = 5)
map(lat~lon, data=geodata, col = type)
map(lat~lon, data=geodata, scaleby = scale, col = type)
}
