% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.R
\name{poly}
\alias{poly}
\title{Compute Polynomial fits of variables}
\usage{
poly(x, ..., degree = 1, raw = TRUE)
}
\arguments{
\item{x}{Variable. The variable you would like to fit a polynomial to.}

\item{degree}{Integer. The degree of the polynomial to fit.}

\item{raw}{Logical. Defaults is \code{TRUE}, meaning the polynomial coefficient outputs can be used in direct calculations/predictions. Setting this value to \code{FALSE} results in \emph{orthogonal} polynomials whose coefficients can't be used to compute predictions by hand.}
}
\description{
Compute Polynomial fits of variables for use in linear models.
}
\examples{
# Use a quadratic polynomial to predict weight based on height.
lm(weight ~ poly(height, 2), data = cdc)

}
