% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{pride}
\alias{pride}
\title{Pride & Prejudice Word Data}
\format{
A data frame with 122204 observations of 1 variable.
}
\usage{
data(pride)
}
\description{
A dataset containing every word in the Pride and Prejudice book.
}
\details{
\itemize{ \item word. The words as they appear in the book, in order. }
}
\keyword{dataset}
