% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{purchases}
\alias{purchases}
\title{Purchases Data (toy dataset)}
\format{
A data frame with 90 observations of 3 variables
}
\usage{
data(purchases)
}
\description{
Toy (fictional) dataset related to purchases to assist in demonstrating various methods of merging data frames.
}
\details{
\itemize{
\item customer_id.
\item number_of_orders. Number of purchases.
\item average_amount_spent. Average amount spent on those purchases.}
}
\keyword{dataset}
