% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_import.R
\name{quick_import}
\alias{quick_import}
\title{Quick Import From Ohmage}
\usage{
quick_import(server = NULL, username = NULL, token = NULL, campaign = NULL)
}
\description{
Quick Import From Ohmage
}
\examples{
quick_import() # Interactively 
quick_import(server = "https://ids.mobilizingcs.org/", username = "guestx", campaign = "urn:class:PD:2017:fall:guestx:ids:p1:foodhabits") # NOT RUN 
quick_import(server = "https://ids.mobilizingcs.org/", username = "guestx", token = token, campaign = "urn:class:PD:2017:fall:guestx:ids:p1:foodhabits") # IF YOU HAVE TOKEN # NOT RUN 

}
