% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflip.R
\name{rflip}
\alias{rflip}
\alias{print.cointoss}
\title{Coin flip.}
\usage{
rflip(n = 1, prob = 0.5, quiet = FALSE, verbose = !quiet)

\method{print}{cointoss}(x, ...)
}
\arguments{
\item{n}{Integer. The number of times to flip the coin.}

\item{prob}{Number between 0 and 1. The probability of getting a "head".
Default value is 0.5.}

\item{quiet}{Logical. Should a message be printed with the output? Default is
\code{FALSE}, meaning a message will be printed.}

\item{verbose}{Logical. The opposite of quiet from above. Should a message be
printed? Default value is \code{TRUE}. Note, either quiet or verbose can be
specified, but not both.}

\item{x}{an object}

\item{\dots}{additional arguments}
}
\description{
Flip a coin 2000 times or fewer.
}
\examples{
# Basic usage. Note the printed output.
rflip()

# The printed output can be turned off using either method:
rflip(quiet = TRUE)
rflip(verbose = FALSE)

# Flip an unfair coin many times
flips <- rflip(100, prob = 0.35)
flips

# The actual output is embedded as an attribute
attr(flips, "sequence")

# Create a bargraph of the output
flips_output <- attr(flips, "sequence")
bargraph(~flips_output)

}
