% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_die.R
\name{roll_die}
\alias{roll_die}
\title{Biased Dice Experiment}
\usage{
roll_die(die, times = 1)
}
\arguments{
\item{die}{\code{"A"} or \code{"B"}. The die you wish to roll.}

\item{times}{Integer. The number of times you'd like to roll the die.}
}
\description{
Roll dice to determine which, if any, is biased.
}
\examples{
# Roll a die once
roll_die("A")

# Roll a die many times then create a histogram.
rolls <- roll_die("A", times = 100)
histogram(rolls, nint = 6)

}
