% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{sentiments}
\alias{sentiments}
\title{Sentiments of Words Data}
\format{
A data frame with 6786 observations of 2 variables
}
\source{
\url{https://github.com/EmilHvitfeldt/textdata/blob/HEAD/R/lexicon_bing.R}
}
\usage{
data(sentiments)
}
\description{
This dataset contains a list of words and estimated positive/negative sentiment based on how they are most often used in the Microsoft Bing search engine.
}
\details{
\itemize{
\item word. Each word that has a defined sentiment.
\item sentiment. Either positive or negative, based on the Bing sentiment lexicon.}
}
\keyword{dataset}
