% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{slasher}
\alias{slasher}
\title{Survival status of actors and actresses in "Slasher" films}
\format{
A data frame with 485 observations of 2 variables
}
\usage{
data(slasher)
}
\description{
A dataset containing survival statuses of actors and actresses in 50 randomly sampled "Slasher" films from 1960 to 2009.
}
\details{
\itemize{ \item gender. the gender of actor/actress \item survival. the survival status of the actor/actress }
}
\keyword{dataset}
