% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{Stop Words Data}
\format{
A data frame with 1149 observations of 1 variable
}
\usage{
data(stop_words)
}
\description{
This dataset contains stop words. Stop words are words that are generally filtered out before or after processing of natural language data because they are insignificant (i.e., a, the, etc.).
}
\details{
\itemize{
\item word.}
}
\keyword{dataset}
