% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwatch.R
\name{stopwatch}
\alias{stopwatch}
\title{Time Perception Stopwatch}
\usage{
stopwatch()
}
\description{
A stopwatch to use for the time perception experiment. To use, simply run the
function \code{stopwatch()} and then hit enter at the desired time to stop
the timer.
}
\examples{
\dontrun{
stopwatch()
}
}
