% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{surveys}
\alias{surveys}
\title{Surveys Data}
\format{
A data frame with 36 observations of 4 variables
}
\usage{
data(surveys)
}
\description{
This dataset contains surveys.
}
\details{
\itemize{
\item customer_id.
\item age.
\item gender.
\item favorite_product.}
}
\keyword{dataset}
