% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail.R
\name{tail}
\alias{tail}
\title{Print last six observations.}
\usage{
tail(x, ..., data = NULL, groups = NULL, na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{Formula or other object to print.}

\item{data}{Data frame. The data where the variables can be found.}
}
\description{
\code{tail} will print the last six values of an object.
}
\examples{
head(~hp, data = mtcars)
tail(~hp, data = mtcars)

}
