% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeuse_format.R
\name{timeuse_format}
\alias{timeuse_format}
\title{Cleans and formats data from the Time Use campaign.}
\usage{
timeuse_format(data)
}
\arguments{
\item{data}{Data frame. The unaltered Time Use campaign data that is
downloaded, uploaded and loaded from the Mobilize landing page.}
}
\value{
The average time spent participating in various activities for each
  participant.
}
\description{
Cleans and formats data from the Time Use campaign. Each observation will be
the average time spent participating in various activities over the data
collection period for each user.
}
\section{How it works}{
 The code cleans and aggregates the data in a 3-step
  process. \enumerate{ \item The first step is to do some preliminary
  cleaning of the data. This includes a step to turn "NOT_DISPLAYED" values
  into 0's. This is done because when a user does not select a category to
  note the amount of time they've participated, it's assumed that the reason
  is because they've  participated in the activity for 0 minutes. \item The
  next step is to sum the amount of time spent doing the various activities
  for each day. This then represents the total amount of time spent
  performing the various activities over the course of each day. \item
  Finally, the third step is to take the summed times for each day from step
  2. and average them over the total number of days they participated in the
  campaign.}
}

\examples{
\dontrun{
timeuse_raw <- read.csv("My class time use data.csv")
timeuse <- timeuse_format(timeuse_raw)
}

}
