% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic Passenger Data}
\format{
A data frame with 1,000 observations of 8 variables
}
\source{
\url{https://hbiostat.org/data/repo/titanic.html}
}
\usage{
data(titanic)
}
\description{
A dataset Containing Information on 1,000 Randomly Sampled Passengers of the
Titanic.
}
\details{
\itemize{ \item name. name of passenger \item age. age of passenger \item
gender. gender of passenger \item fare. amount paid for ticket (in Pounds) to sail on the
Titanic \item class. the class hospitality/room aboard the Titanic \item
embarked. location passenger first stepped on board the Titanic \item
survived. the survival status of the passenger \item titled. Does the person have a title other than Mr., Mrs., Miss., etc. in their name or not}
}
\seealso{
\link{titanic_test}
}
\keyword{dataset}
