% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{transfer_download}
\alias{transfer_download}
\title{transfer_download}
\usage{
transfer_download(url = NULL, passcode = NULL, destfile = NULL)
}
\arguments{
\item{url}{A character string naming the URL of a resource to be downloaded}

\item{passcode}{A character string provided by transfer_upload()}

\item{destfile}{A character string with the name where the downloaded file is saved. Tilde-expansion is performed.}
}
\description{
Share files - Download via transfer.sh
}
\examples{

transfer_download('do.R', passcode='ayU4v') # NOT RUN
transfer_download('ayU4v/do.R') # NOT RUN
transfer_download('do.R', passcode='ayU4v', destfile = 'do.R.copy') # NOT RUN
transfer_download('ayU4v/do.R', destfile = 'do.R.copy') # NOT RUN

}
