% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{transfer_upload}
\alias{transfer_upload}
\title{transfer_upload}
\usage{
transfer_upload(file = NULL)
}
\arguments{
\item{file}{path to an existing file on disk}
}
\description{
Share files - Upload via transfer.sh
}
\examples{

transfer_upload("my.csv") # NOT RUN

}
