% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplot.R
\name{treeplot}
\alias{treeplot}
\title{Tree Plot for Tree-based Models}
\usage{
treeplot(model, ...)
}
\arguments{
\item{model}{The tree-based model to plot.}

\item{...}{Additional plotting parameters that can be passed. See help
documentation for rpart.plot for additional information}
}
\description{
Creates a visualization of the tree created when using a tree-based model.
}
\examples{
m <- tree(survived ~ age + gender + class, data = titanic)
treeplot(m)

}
