% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{wordcloud}
\alias{wordcloud}
\title{Word Clouds}
\usage{
wordcloud(formula, data, max.words = 100, color = "black")
}
\arguments{
\item{data}{data frame. The data where the variables can be found.}

\item{max.words}{argument. An argument to designate a different max number of words to plot in the word cloud (default is 100).}

\item{color}{argument. An argument to designate the color of the word cloud.}

\item{x}{\code{formula}. An expression of the form \code{y ~ x} where \code{x} is the words we are selecting and \code{y} is the frequency with which the word occurs. Note that \code{x} and \code{y} are the variable names where these are located.}
}
\description{
Create a word cloud.
}
\section{Note}{

For additional options for creating word clouds, look at the help file found in the \code{wordcloud} package or run \code{help(wordcloud, package="wordcloud")} in the console. ALSO: note in the examples that data must be formatted a particular way in order to use the wordcloud function.
}

\examples{
data(pride)
pride <- tally(~word, data = pride, format = "data.frame")
wordcloud(Freq ~ word, data = pride)

}
