% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot.R
\name{xyplot}
\alias{xyplot}
\title{Scatter Plots}
\usage{
xyplot(x, data, ...)
}
\arguments{
\item{x}{\code{formula}. An expression of the form \code{y ~ x | z} where \code{x} is the variable we want on the x-axis, \code{y} is the variable we want on the y-axis and \code{z} is an optional variable we would like to use for faceting.}

\item{data}{data frame. The data where the variables can be found.}

\item{groups}{factor. An additional variable to use for coloring the individual points in the plot.}
}
\description{
Create a scatter plot.
}
\section{Note}{

For additional options for creating bargraphs, look at the help file found in the \code{lattice} package or run \code{help(xyplot, package="lattice")} in the console.
}

\examples{
data(cdc)
xyplot(weight~height, data = cdc)
xyplot(weight ~ height | gender, data = cdc)
xyplot(weight ~ height, data = cdc, groups = gender)

}
